+++ tests/test_roundtrip.c
@@ -42,7 +42,7 @@ static char* test_roundtrip(caterva_context_t *ctx, uint8_t itemsize, uint8_t nd
     /* Create original data */
     size_t buffersize = itemsize;
     for (int i = 0; i < ndim; ++i) {
-        buffersize *= shape[i];
+        buffersize *= (size_t) shape[i];
     }
     uint8_t *buffer = malloc(buffersize);
     MU_ASSERT("Buffer filled incorrectly", fill_buf(buffer, itemsize, buffersize / itemsize));
